requirements.txt
django
channels[daphne]
channels_redis
websocket-client
rel
setuptools
djangorestframework
python-dotenv
colorlog
django-split-settings
humanfriendly
django-health-check
